Change the estructure of Datebase
Change TreeMaps to HashSet for problems of compatibility with the Database, and fixed the DatabaseInitializer.

+++ Daw/src/main/java/com/urjc/daw/Models/Concept/ConceptRepository.java
@@ -9,6 +9,6 @@ import org.springframework.data.repository.CrudRepository;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface ConceptRepository extends JpaRepository<Concept, Integer> {
+public interface ConceptRepository extends JpaRepository<Concept, Long> {
 
 }
